#25 Added helper classes and methods for centralized exception management
- Throw: changed from internal to public static helper class to simplify throwing exceptions under various conditions.
- ExceptionActivator: changed from internal to public static helper class for dynamically creating exception instances.
- ExceptionActivator.Create method: changed from internal to public generic method to create exception instances of a specified type.
- Improved code readability and maintainability.
- Added XML documentation comments to relevant methods.
