+++ guard/src/Guards/NullOrEmpty/Is.NullOrEmpty.cs
@@ -1,6 +1,5 @@
 using System.Collections;
 using System.Linq.Expressions;
-using SGuard.Visitor;
 
 namespace SGuard;
