+++ guard/src/Guards/NullOrEmpty/ThrowIf.NullOrEmpty.cs
@@ -2,7 +2,6 @@ using System.Collections.Concurrent;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq.Expressions;
 using SGuard.Exceptions;
-using SGuard.Visitor;
 
 namespace SGuard;
